16,32,64 bit

+++ Assets/Scripts/Simulation/SimPin.cs
@@ -68,8 +68,8 @@ namespace DLS.Simulation
 
 				for (int i = 0; i < State.BitCount; i++)
 				{
-					uint targetLogicLevel = source.State.GetBit(i);
-					uint currentLogicLevel = State.GetBit(i);
+					UInt64 targetLogicLevel = source.State.GetBit(i);
+					UInt64 currentLogicLevel = State.GetBit(i);
 
 					// Ignore disconnected input. Always accept input if current state is disconnected. Otherwise, choose randomly.
 					if (targetLogicLevel != PinState.LogicDisconnected && (currentLogicLevel == PinState.LogicDisconnected || acceptConflictingInput))
